feat: create ChartRender component to refactor render chart
