+++ packages/ui/src/components/templates/Dashboard/Chart/ChartRender.tsx
@@ -1,0 +1,93 @@
+/*
+ * Axelor Business Solutions
+ *
+ * Copyright (C) 2024 Axelor (<http://axelor.com>).
+ *
+ * This program is free software: you can redistribute it and/or  modify
+ * it under the terms of the GNU Affero General Public License, version 3,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import React from 'react';
+import {Dimensions} from 'react-native';
+import Chart from '../chart-type';
+import BarChart from './BarChart';
+import LineChart from './LineChart';
+import PieChart from './PieChart';
+
+const getGraphWidth = (nbGraphInLine: number) => {
+  if (nbGraphInLine === 1) {
+    return Dimensions.get('window').width;
+  } else if (Dimensions.get('window').width < 500) {
+    return Dimensions.get('window').width / 2;
+  } else {
+    return Dimensions.get('window').width / nbGraphInLine;
+  }
+};
+
+const BarChartRender = (datasets, key, title, widthGraph) => {
+  return (
+    <BarChart
+      key={key}
+      title={title}
+      datasets={datasets}
+      widthGraph={widthGraph}
+    />
+  );
+};
+
+const LineChartRender = (datasets, key, title, widthGraph) => {
+  return (
+    <LineChart
+      key={key}
+      title={title}
+      datasets={datasets}
+      widthGraph={widthGraph}
+    />
+  );
+};
+
+const PieChartRender = (datasets, key, title, widthGraph, type) => {
+  return (
+    <PieChart
+      key={key}
+      title={title}
+      datasets={datasets[0]}
+      widthGraph={widthGraph}
+      donut={type === Chart.chartType.donut}
+    />
+  );
+};
+
+const ChartRender = ({graph, indexGraph = 0, nbGraphInLine = null}) => {
+  const {dataList, title, type} = graph;
+  const widthGraph =
+    nbGraphInLine != null
+      ? getGraphWidth(nbGraphInLine)
+      : Dimensions.get('window').width;
+
+  switch (type) {
+    case Chart.chartType.bar:
+      return BarChartRender(dataList, indexGraph, title, widthGraph);
+
+    case Chart.chartType.pie:
+    case Chart.chartType.donut:
+      return PieChartRender(dataList, indexGraph, title, widthGraph, type);
+
+    case Chart.chartType.line:
+      return LineChartRender(dataList, indexGraph, title, widthGraph);
+
+    default:
+      return null;
+  }
+};
+
+export default ChartRender;
