+++ packages/ui/src/components/templates/Dashboard/Dashboard.tsx
@@ -17,14 +17,14 @@
  */
 
 import React from 'react';
-import {Dimensions, StyleSheet, View} from 'react-native';
+import {StyleSheet, View} from 'react-native';
 import {ScrollView} from '../../atoms';
 import Chart from './chart-type';
 import {
   Data,
   Max_Number_Graph_Line as MAX_GRAPH_PER_LINE,
 } from './dashboard.helper';
-import {BarChart, LineChart, PieChart} from './Chart';
+import ChartRender from './Chart/ChartRender';
 
 interface Graph {
   type: keyof typeof Chart.chartType;
@@ -41,70 +41,6 @@ interface DashboardProps {
   lineList: Line[];
 }
 
-const getGraphWidth = (nbGraphInLine: number) => {
-  if (nbGraphInLine === 1) {
-    return Dimensions.get('window').width;
-  } else if (Dimensions.get('window').width < 500) {
-    return Dimensions.get('window').width / 2;
-  } else {
-    return Dimensions.get('window').width / nbGraphInLine;
-  }
-};
-
-const BarChartRender = (datasets, key, title, widthGraph) => {
-  return (
-    <BarChart
-      key={key}
-      title={title}
-      datasets={datasets}
-      widthGraph={widthGraph}
-    />
-  );
-};
-
-const LineChartRender = (datasets, key, title, widthGraph) => {
-  return (
-    <LineChart
-      key={key}
-      title={title}
-      datasets={datasets}
-      widthGraph={widthGraph}
-    />
-  );
-};
-
-const PieChartRender = (datasets, key, title, widthGraph, type) => {
-  return (
-    <PieChart
-      key={key}
-      title={title}
-      datasets={datasets[0]}
-      widthGraph={widthGraph}
-      donut={type === Chart.chartType.donut}
-    />
-  );
-};
-
-const renderChart = (graph, indexGraph, nbGraphInLine) => {
-  const {dataList, title, type} = graph;
-  const widthGraph = getGraphWidth(nbGraphInLine);
-
-  switch (type) {
-    case Chart.chartType.bar:
-      return BarChartRender(dataList, indexGraph, title, widthGraph);
-
-    case Chart.chartType.pie:
-    case Chart.chartType.donut:
-      return PieChartRender(dataList, indexGraph, title, widthGraph, type);
-
-    case Chart.chartType.line:
-      return LineChartRender(dataList, indexGraph, title, widthGraph);
-
-    default:
-      return null;
-  }
-};
-
 const Dashboard = ({style, lineList}: DashboardProps) => {
   return (
     <ScrollView style={[styles.container, style]}>
@@ -120,7 +56,14 @@ const Dashboard = ({style, lineList}: DashboardProps) => {
           <View style={styles.lineContainer} key={indexLine}>
             {limitedGraphList?.map((graph, indexGraph) => {
               if (graph.dataList?.[0]?.length > 0) {
-                return renderChart(graph, indexGraph, nbGraphInLine);
+                return (
+                  <ChartRender
+                    key={indexGraph}
+                    graph={graph}
+                    indexGraph={indexGraph}
+                    nbGraphInLine={nbGraphInLine}
+                  />
+                );
               }
               return null;
             })}
